+++ changelog.md
@@ -19,6 +19,8 @@ This version is not yet released. If you are reading this on the website, then t
 ### Interpreter
 - Numbers that seem to have a floating-point epsilon rounding error will be output with the epsilon noted
 - When hovering a user-defined function's name, the language server will now show whether it is compatible with [`un °`](https://uiua.org/docs/un) and [`under ⍜`](https://uiua.org/docs/under)
+### Website
+- Add [Tutorial Introduction](https://uiua.org/docs/introduction)
 
 ## 0.8.0 - 2024-01-31
 ### Language